Refactor(package) : 축약 제거
