+++ src/main/java/chess/domain/strategy/PieceStrategy.java
@@ -1,6 +1,6 @@
 package chess.domain.strategy;
 
-import chess.domain.dto.req.MoveRequest;
+import chess.domain.dto.request.MoveRequest;
 
 public interface PieceStrategy {
