+++ src/main/java/chess/domain/strategy/bishop/BishopStrategy.java
@@ -1,7 +1,7 @@
 package chess.domain.strategy.bishop;
 
 import chess.domain.dto.PositionDto;
-import chess.domain.dto.req.MoveRequest;
+import chess.domain.dto.request.MoveRequest;
 import chess.domain.strategy.PieceStrategy;
 
 public class BishopStrategy implements PieceStrategy {
