+++ src/test/java/chess/domain/strategy/knight/KnightStrategyTest.java
@@ -3,7 +3,7 @@ package chess.domain.strategy.knight;
 import chess.domain.Color;
 import chess.domain.Position;
 import chess.domain.dto.PositionDto;
-import chess.domain.dto.req.MoveRequest;
+import chess.domain.dto.request.MoveRequest;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
