+++ src/test/java/chess/domain/strategy/pawn/PawnStrategyTest.java
@@ -2,7 +2,7 @@ package chess.domain.strategy.pawn;
 
 import chess.domain.Position;
 import chess.domain.dto.PositionDto;
-import chess.domain.dto.req.MoveRequest;
+import chess.domain.dto.request.MoveRequest;
 import chess.domain.strategy.PieceStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
