+++ src/test/java/chess/domain/strategy/rook/RookStrategyTest.java
@@ -3,7 +3,7 @@ package chess.domain.strategy.rook;
 import chess.domain.Color;
 import chess.domain.Position;
 import chess.domain.dto.PositionDto;
-import chess.domain.dto.req.MoveRequest;
+import chess.domain.dto.request.MoveRequest;
 import chess.domain.strategy.PieceStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
